Move sharable code to main module

+++ src/commonMain/kotlin/sh/uffle/koms/ByteListExtensions.kt
@@ -1,3 +1,3 @@
 package sh.uffle.koms
 
-internal fun List<Byte>.asInt() = takeLast(4).fold(0) { result, value -> result.shl(8).or(value.toInt()) }
+internal fun List<Byte>.asInt() = takeLast(4).fold(0) { result, value -> result.shl(8).or(value.toUByte().toInt()) }
